LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/text-underline-position-expected.txt


PASS Can set 'text-underline-position' to CSS-wide keywords
FAIL Can set 'text-underline-position' to var() references assert_equals: expected 2 but got 1
PASS Can set 'text-underline-position' to the 'auto' keyword
PASS Can set 'text-underline-position' to the 'under' keyword
FAIL Can set 'text-underline-position' to the 'left' keyword Invalid values
FAIL Can set 'text-underline-position' to the 'right' keyword Invalid values
PASS Setting 'text-underline-position' to a length throws TypeError
PASS Setting 'text-underline-position' to a percent throws TypeError
PASS Setting 'text-underline-position' to a time throws TypeError
PASS Setting 'text-underline-position' to an angle throws TypeError
PASS Setting 'text-underline-position' to a flexible length throws TypeError
PASS Setting 'text-underline-position' to a number throws TypeError
PASS Setting 'text-underline-position' to a URL throws TypeError
PASS Setting 'text-underline-position' to a transform throws TypeError
FAIL 'text-underline-position' does not supported 'under left' assert_not_equals: Unsupported value must not be null got disallowed value null
FAIL 'text-underline-position' does not supported 'right under' assert_not_equals: Unsupported value must not be null got disallowed value null
